adding leapmotion js support

+++ molecularcontroltoolkit/src/org/odonoghuelab/molecularcontroltoolkit/GestureListener.java
@@ -13,14 +13,14 @@ public interface GestureListener {
 	 * Zoom action has been performed
 	 * @param zoom relative amount to zoom
 	 */
-	public abstract void triggerZoom(int zoom);
+	public abstract void triggerZoom(float zoom);
 
 	/** 
 	 * Pan action has been performed
 	 * @param panX relative amount to pan in X 
 	 * @param panY relative amount to pan in Y 
 	 */
-	public abstract void triggerPan(int panX, int panY);
+	public abstract void triggerPan(float panX, float panY);
 
 	/**
 	 * Rotate action has been performed
@@ -28,7 +28,7 @@ public interface GestureListener {
 	 * @param rotateY relative amount to rotate in Y 
  	 * @param rotateZ relative amount to rotate in Z 
 	 */
-	public abstract void triggerRotate(int rotateX, int rotateY, int rotateZ);
+	public abstract void triggerRotate(float rotateX, float rotateY, float rotateZ);
 
 	/**
 	 * Reset the view
@@ -42,8 +42,8 @@ public interface GestureListener {
 	
 	/**
 	 * Point the mouse cursor at the given x,y location
-	 * @param x
-	 * @param y
+	 * @param x x location
+	 * @param y y location
 	 */
 	public abstract void point(float x, float y);
